requirements.txt
psycopg2-binary==2.9.3
asgiref==3.5.2
Django==4.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
Unipath==1.1
whitenoise
python-dotenv
pyyaml
uritemplate